Tests/LibWeb/Text/expected/css/stylesheet-add-remove-host-targeted-invalidation-shadow-root.txt

before add: rgb(0, 0, 0)
after add: rgb(255, 0, 0)
after remove: rgb(0, 0, 0)
